+++ examples/tv-app/android/java/src/com/matter/tv/server/tvapp/MessagesManager.java
@@ -28,7 +28,7 @@ public interface MessagesManager {
       int priority,
       int messageControl,
       long startTime,
-      int duration,
+      long duration,
       String messageText,
       HashMap<Long, String> responseOptions);
